refactor: move forRange to its own file
