+++ src/game-logic/reducers/forRange.test.js
@@ -1,0 +1,32 @@
+import { testCrop } from '../../test-utils'
+
+import { waterPlot } from './waterPlot'
+import { forRange } from './forRange'
+
+describe('forRange', () => {
+  test('calls given reducer on range of plots', () => {
+    const { field } = forRange(
+      {
+        field: [
+          [
+            testCrop({ itemId: 'sample-crop-1' }),
+            testCrop({ itemId: 'sample-crop-1' }),
+          ],
+          [testCrop({ itemId: 'sample-crop-1' })],
+          [],
+          [],
+          [testCrop({ itemId: 'sample-crop-1' })],
+        ],
+      },
+      waterPlot,
+      1,
+      1,
+      1
+    )
+
+    expect(field[0][0].wasWateredToday).toBe(true)
+    expect(field[0][1].wasWateredToday).toBe(true)
+    expect(field[1][0].wasWateredToday).toBe(true)
+    expect(field[4][0].wasWateredToday).toBe(false)
+  })
+})
