+++ src/game-logic/reducers/index.js
@@ -68,37 +68,7 @@ export * from './changeCowName'
 export * from './selectCow'
 export * from './updateAchievements'
 export * from './adjustLoan'
-
-/**
- * @param {farmhand.state} state
- * @param {number} adjustmentAmount This should be a negative number if the
- * @param {number} rangeRadius
- * @param {number} x
- * @param {number} y
- * @param {...any} args Passed to fieldFn.
- * @returns {farmhand.state}
- */
-export const forRange = (
-  state,
-  fieldFn,
-  rangeRadius,
-  plotX,
-  plotY,
-  ...args
-) => {
-  const startX = Math.max(plotX - rangeRadius, 0)
-  const endX = Math.min(plotX + rangeRadius, state.field[0].length - 1)
-  const startY = Math.max(plotY - rangeRadius, 0)
-  const endY = Math.min(plotY + rangeRadius, state.field.length - 1)
-
-  for (let y = startY; y <= endY; y++) {
-    for (let x = startX; x <= endX; x++) {
-      state = fieldFn(state, x, y, ...args)
-    }
-  }
-
-  return state
-}
+export * from './forRange'
 
 /**
  * @param {farmhand.state} state
